Moved all game controller logic to separate package
